requirements.txt
bcrypt==5.0.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
